requirements.txt
Flask==2.3.*
cfenv==0.5.3
sap-xssec==4.*